Phase-2/requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-RESTX==0.5.1
SQLAlchemy==1.4.23
pyodbc==4.0.32
python-dotenv==0.19.0
PyJWT==2.3.0
Werkzeug==2.0.1
python-dateutil==2.8.2
email-validator==1.1.3
cryptography==3.4.7
requests==2.26.0
pytest==6.2.5
pytest-cov==2.12.1
black==21.7b0
flake8==3.9.2
mypy==0.910
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
python-multipart==0.0.5
aiofiles==0.7.0
python-magic==0.4.24
python-json-logger==2.0.2
celery==5.3.6
redis==3.5.3
pytest-flask==1.2.0
pytest-mock==3.6.1
pytest-asyncio==0.15.1
coverage==5.5
flower==1.0.0
prometheus-client==0.11.0
sentry-sdk[flask]==1.3.1
flask-cors==3.0.10
flask-limiter==1.4 